services/machine_learning/requirements.txt
numpy
scipy
scikit-learn
jupyter
pandas
sklearn
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.4
Flask-Testing==0.6.2
gunicorn==19.8.1
flask-cors==3.0.6
PyJWT==1.7.1
pytest==4.1.1
pytest-cov==2.6.1
requests==2.21.0
Flask-Migrate==2.4.0
Flask-Script==2.0.6
